refactor(module): move files to src folder, convert style of every file
